updated files
Added code for comments and made a form. I downloaded Bootstrap and JQuery and don't use cdn's anymore.
I added regex to the js-file and in the html-file I added a modal and an alert  for the form . For the comments I used media objects. Furthermore I changed the width of the carousel, now it's equal to the screensize.